Use dynamic imports for noble (#22)
Horribly enough, importing noble will fail on Linux if bluetooth is not
supported. To work around that and have a chance to handle the exception
gracefully, dynamically import noble but continue to import noble types
eagerly. Using noble doesn’t get better but at least it’s all contained
in a single module.